CharEvaluate/CharFeature/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.6.0
lightgbm>=3.2.0
scikit-learn>=0.24.0
PyYAML>=5.4.0
pandas>=1.2.0
tqdm>=4.60.0
detectron2@git+https://github.com/facebookresearch/detectron2.git
